requirements.txt
asgiref
certifi==2020.12.5
cffi
cryptography
defusedxml==0.7.1
diff-match-patch==20200713
Django @ file:///tmp/build/80754af9/django_1613762073373/work
django-autocomplete-light==3.8.2
django-ckeditor==6.0.0
django-crispy-forms==1.11.1
django-formtools==2.2
django-import-export==2.5.0
django-js-asset==1.2.2
django-rest-framework==0.1.0
django-reversion==3.0.9
djangorestframework==3.12.2
et-xmlfile==1.0.1
future==0.18.2
gunicorn==20.0.4
httplib2==0.19.0
MarkupPy==1.14
mistune==0.8.4
odfpy==1.4.1
openpyxl==3.0.7
psycopg2 @ file:///C:/ci/psycopg2_1612298619396/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyMySQL @ file:///C:/ci/pymysql_1610464866457/work
pyparsing==2.4.7
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==5.4.1
six @ file:///C:/ci/six_1605187303045/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tablib==3.0.0
wincertstore==0.2
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
